Extended the models file. Minor fixes
Extended the models file with the `ThumbnailAbstract` that implements the
upload of small images (800x800 by default).
The function `crop_image`, binded on post-save signal of models that
extends the `ThumbnailAbstract`, correctly resize and crop larger images
to the given size.
Stubs for `lattice/settings.py` (App sepcific settings) and `requirements.txt`